Readded ownership requirements to patch and delete
These requirements were removed early to make development easier.

+++ app/routes/combat_routes.js
@@ -6,7 +6,7 @@ const Combat = require('../models/combat')
 const customErrors = require('../../lib/custom_errors')
 
 const handle404 = customErrors.handle404
-// const requireOwnership = customErrors.requireOwnership
+const requireOwnership = customErrors.requireOwnership
 
 const removeBlanks = require('../../lib/remove_blank_fields')
 const requireToken = passport.authenticate('bearer', { session: false })
@@ -43,13 +43,13 @@ router.post('/combats', requireToken, (req, res, next) => {
 })
 
 // UPDATE
-router.patch('/combats/:id', removeBlanks, (req, res, next) => {
+router.patch('/combats/:id', requireToken, removeBlanks, (req, res, next) => {
   delete req.body.combat.owner
 
   Combat.findById(req.params.id)
     .then(handle404)
     .then(combat => {
-      // requireOwnership(req, combat)
+      requireOwnership(req, combat)
       return combat.updateOne(req.body.combat)
     })
     .then(() => res.sendStatus(204))
@@ -57,11 +57,11 @@ router.patch('/combats/:id', removeBlanks, (req, res, next) => {
 })
 
 // DESTROY
-router.delete('/combats/:id', (req, res, next) => {
+router.delete('/combats/:id', requireToken, (req, res, next) => {
   Combat.findById(req.params.id)
     .then(handle404)
     .then(combat => {
-      // requireOwnership(req, combat)
+      requireOwnership(req, combat)
       combat.deleteOne()
     })
     .then(() => res.sendStatus(204))
